+++ wolframclient/evaluation/__init__.py
@@ -2,32 +2,28 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
-from wolframclient.evaluation.cloud import SecuredAuthenticationKey, UserIDPassword, WolframAPICall, WolframCloudSession, WolframCloudSessionAsync, WolframServer
-from wolframclient.evaluation.result import WolframAPIResponse, WolframEvaluationJSONResponse, WolframKernelEvaluationResult, WolframResult
+from wolframclient.evaluation.cloud import (
+    SecuredAuthenticationKey, UserIDPassword, 
+    WolframAPICall, WolframCloudSession, WolframCloudSessionFuture,
+    WolframAPICallAsync, WolframCloudAsyncSession, 
+    WolframServer)
+from wolframclient.evaluation.result import (
+    WolframAPIResponse, WolframAPIResponseAsync, 
+    WolframEvaluationJSONResponse, WolframEvaluationJSONResponseAsync,
+    WolframKernelEvaluationResult, 
+    WolframResult, WolframResultBase)
 from wolframclient.utils import six
 
-if six.JYTHON:
-    __all__ = [
-        'WolframServer', 'WolframCloudSession', 'WolframCloudSessionAsync',
-        'WolframAPICall', 'SecuredAuthenticationKey', 'UserIDPassword',
-        'WolframResult', 'WolframKernelEvaluationResult', 'WolframAPIResponse',
-        'WolframEvaluationJSONResponse'
-    ]
-else:
-    from wolframclient.evaluation.kernel import WolframLanguageSession, WolframLanguageFutureSession
-    __all__ = [
-        'WolframAPICall', 'WolframServer', 'WolframCloudSession',
-        'WolframCloudSessionAsync', 'WolframAPICall',
-        'SecuredAuthenticationKey', 'UserIDPassword', 'WolframLanguageSession',
-        'WolframLanguageFutureSession', 'WolframResult',
-        'WolframKernelEvaluationResult', 'WolframAPIResponse',
-        'WolframEvaluationJSONResponse'
-    ]
-    try:
-        from wolframclient.evaluation.kernel import WolframLanguageAsyncSession
-        from wolframclient.evaluation.kernel import WolframKernelPool, parallel_evaluate
-        __all__.append(WolframKernelPool.__name__)
-        __all__.append(WolframLanguageAsyncSession.__name__)
-        __all__.append(parallel_evaluate.__name__)
-    except:
-        pass
+from wolframclient.evaluation.kernel import WolframLanguageSession, WolframLanguageFutureSession
+from wolframclient.evaluation.kernel import WolframLanguageAsyncSession
+from wolframclient.evaluation.kernel import WolframKernelPool, parallel_evaluate
+
+__all__ = [
+    'WolframAPICall', 'WolframServer', 'WolframCloudSession',
+    'WolframCloudAsyncSession', 'WolframAPICall',
+    'SecuredAuthenticationKey', 'UserIDPassword', 'WolframLanguageSession',
+    'WolframLanguageFutureSession', 'WolframResult',
+    'WolframKernelEvaluationResult', 'WolframAPIResponse',
+    'WolframEvaluationJSONResponse',
+    'WolframLanguageAsyncSession', 'WolframKernelPool', 'parallel_evaluate'
+]