+++ wolframclient/evaluation/cloud/__init__.py
@@ -2,12 +2,14 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
-from wolframclient.evaluation.cloud.cloudsession import WolframAPICall, WolframCloudSession, WolframCloudSessionAsync
+from wolframclient.evaluation.cloud.cloudsession import WolframAPICall, WolframCloudSession, WolframCloudSessionFuture
+from wolframclient.evaluation.cloud.asynccloudsession import WolframAPICallAsync, WolframCloudAsyncSession
 from wolframclient.evaluation.cloud.base import SecuredAuthenticationKey, UserIDPassword
 from wolframclient.evaluation.cloud.server import WolframServer
 
 
 __all__ = [
-    'WolframServer', 'WolframCloudSession', 'WolframCloudSessionAsync',
-    'WolframAPICall', 'SecuredAuthenticationKey', 'UserIDPassword'
+    'WolframServer', 'WolframCloudSession', 'WolframCloudSessionFuture', 'WolframAPICall', 
+    'SecuredAuthenticationKey', 'UserIDPassword',
+    'WolframAPICallAsync', 'WolframCloudAsyncSession'
 ]