+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Major refactoring of the evaluation module.
     - add base classes to define evaluation schemes, synchronous and asynchronous.
     - add cloud session implementation based on asyncio and aiohttp.
+    - rename asynchronous session based on future module from `WolframCloudAsyncSession` to `WolframCloudSessionFuture`.
+    - add new implementation of `WolframCloudAsyncSession` based on asyncio and coroutines.
 - Cloud session named option `authentication` renamed `credentials`.